Fixes on map
We have no map-data outside Denmark, so the go-methods cannot go further than to the edges.
The zoom-rect-function can now be activated by drawing any rectangle instead of only from top left to bottom right.
We have to fix the case when a user incidentally clicks the map, which results in a factor-0-error.