Lots of changes, see list:
Create Common-roadType-enum.
Make GroupE-map work with the new enum.
Parse Krak-data to common .csv-format.
Optimization when determining a roadtype. Before we ran a loop over all types, now we look it up in a map.
As some of the data-sets are empty when using one format, these sets should not be used for drawing data from the other format.
KDTree-search-limit-change to a much lower value.
BasicStrokes are now saved in a map for many values, to not create new Strokes every time the map is drawn.
Add code to controller that calculates the initial loading and drawing time.
Limit the factor to at least 0.5, this means that you can only zoom this far.
Refactor Edge and Node-classes to use private fields with Accessor-methods.
Make a separate package for parsing-classes (and old Edge and Node-classes).
Street numbers and information about city, postcode and so on, is not in the common-street-class now. (This should be fixed and reverted)
Parse OSM to common .csv-format. (Check KrakParser.java or edges.csv for the list of elements in the right order) (see above)
Make the user choose which dataset to use.
Try to fix the Rightclick-move-bug.